requirements.txt
Django<=4.2.3
djangorestframework<=3.12.4
django-filter<=2.4.0
django-cors-headers<=3.7.0
joblib<=1.2.0
Pillow
django-phonenumber-field<=5.2.0
phonenumbers<=8.12.31
djangorestframework-simplejwt<=4.8.0
requests<=2.26.0
mysql<=0.0.3
mysql-connector-python<=8.0.32
mysqlclient<=2.1.1
requests<=2.28.0
urllib3<=1.26.9
djangorestframework-simplejwt<=4.10.0

